+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,14 +1,3 @@
 package blackjack.domain
 
-data class Card(val mark: CardMark, val value: CardValue) {
-
-    companion object {
-        private val CARDS: List<Card> =
-            CardMark.values().map { mark ->
-                CardValue.values().map { value ->
-                    Card(mark, value)
-                }
-            }.flatten()
-        fun all(): List<Card> = CARDS
-    }
-}
+data class Card(val mark: CardMark, val value: CardValue)
