+++ src/test/kotlin/blackjack/domain/BlackJackTest.kt
@@ -8,7 +8,7 @@ class BlackJackTest {
     fun `게임 결과를 반환한다`() {
         val blackJack = BlackJack(
             participants = Participants(Dealer(), listOf(Guest("아크"), Guest("로피"))),
-            cardDeck = CardDeck(Card.all()),
+            cardDeck = CardDeck(Cards.all()),
         ).apply {
             dealer.draw(Card(CardMark.CLOVER, CardValue.QUEEN))
             dealer.draw(Card(CardMark.CLOVER, CardValue.QUEEN))
