+++ src/test/kotlin/blackjack/domain/CardDeckTest.kt
@@ -8,13 +8,13 @@ class CardDeckTest {
 
     @Test
     fun `카드덱 잘 있는지 확인`() {
-        val cardDeck = CardDeck(Card.all())
+        val cardDeck = CardDeck(Cards.all())
         assertThat(cardDeck.size).isEqualTo(52)
     }
 
     @Test
     fun `카드 draw가 잘 되었는지 확인`() {
-        val cardDeck = CardDeck(Card.all())
+        val cardDeck = CardDeck(Cards.all())
         cardDeck.nextCard()
         assertThat(cardDeck.size).isEqualTo(51)
     }
@@ -22,14 +22,14 @@ class CardDeckTest {
     @Test
     fun `초기 카드는 52장이다`() {
         assertThrows<IllegalArgumentException> {
-            CardDeck(Card.all().take(51))
+            CardDeck(Cards.all().take(51))
         }
     }
 
     @Test
     fun `카드는 중복이 없어야 합니다`() {
         assertThrows<IllegalArgumentException> {
-            CardDeck(Card.all() + Card.all())
+            CardDeck(Cards.all() + Cards.all())
         }
     }
 }
